WhiteList Email module - listing, edit, delete, filter and pagination and Table empty row fixes, Max 5 user alias add at once

+++ app/src/Controller/Api/OptIn/OptInEmailController.php
@@ -5,7 +5,6 @@ namespace App\Controller\Api\OptIn;
 use App\Domain\Entity\OptIn\OptInEmail\OptInEmail;
 use App\Domain\Entity\OptIn\OptInEmail\OptInEmailRepository;
 use App\Messenger\Validation;
-use IteratorAggregate;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -13,6 +12,8 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
+use OutOfBoundsException;
+use IteratorAggregate;
 
 #[Route('/api/opt-in/emails')]
 class OptInEmailController
@@ -21,22 +22,25 @@ class OptInEmailController
     #[Route('', methods: ['GET'])]
     #[IsGranted('OPTIN_EMAIL_LIST')]
     public function list(
-        Request               $request,
+        Request              $request,
         OptInEmailRepository $optInEmailRepository
     ): Response
     {
+        $query = $request->query->get('query');
         $start = $request->query->get('start');
         $max = $request->query->get('max') ?? 20;
-        $emails = $optInEmailRepository->findAll($start, $max);
+        $sortBy = $request->query->get('sortBy');
+        $desc = $request->query->get('desc');
+        $emails = $optInEmailRepository->findAll($query, $start, $max, $sortBy, boolval($desc));
 
         $count = is_array($emails) ? count($emails) : $emails->count();
 
         if ($emails instanceof IteratorAggregate) {
-            $emails = $emails->getIterator();
+            $emails = (array)$emails->getIterator();
         }
 
         return new JsonResponse([
-            'results' => $emails,
+            'results' => $count === 0 ? [] : $emails,
             'count' => $count,
         ]);
     }
@@ -44,36 +48,51 @@ class OptInEmailController
     #[Route('', methods: ['POST'])]
     #[IsGranted('OPTIN_EMAIL_CREATE')]
     public function create(
-        Request               $request,
-        ValidatorInterface    $validator,
+        Request              $request,
+        ValidatorInterface   $validator,
         OptInEmailRepository $optInEmailRepository
     ): Response
     {
         $body = $request->getContent();
         $data = json_decode($body, true);
-        $optinEmail = OptInEmail::create($data['email'] ?? '');
-        $errors = $validator->validate($optinEmail);
-        if (count($errors) > 0) {
-            return Validation::getViolations($errors);
+
+        $data['email'] = is_array($data['email']) ? $data['email'] : array($data['email']);
+
+        $emails = [];
+        foreach ($data['email'] as $email) {
+            $emailToCreate = new OptInEmail($email);
+            $errors = $validator->validate($emailToCreate);
+
+            if (count($errors) > 0) {
+                return Validation::getViolations($errors);
+            }
+            $emails[] = $emailToCreate;
         }
 
-        $optinEmail = $optInEmailRepository->save($optinEmail);
-        $params = ['email' => $optinEmail->getEmail()];
-        return new JsonResponse($params, Response::HTTP_CREATED);
+        foreach ($emails as $email) {
+            $optInEmailRepository->save($email);
+        }
+
+        return new JsonResponse('Data has been added successfully!');
     }
 
-    #[Route('/{optInEmail}', methods: ['PUT'])]
-    #[IsGranted('OPTIN_EMAIL_EDIT', subject: 'optInEmail')]
-    #[ParamConverter('optInEmail', class: OptInEmail::class, converter: 'app.optInEmail')]
+    #[Route('/edit', methods: ['PUT'])]
+    #[IsGranted('OPTIN_EMAIL_EDIT')]
     public function edit(
-        OptInEmail           $optInEmail,
-        Request               $request,
-        ValidatorInterface    $validator,
+        Request              $request,
+        ValidatorInterface   $validator,
         OptInEmailRepository $optInEmailRepository
     ): Response
     {
         $body = $request->getContent();
         $data = json_decode($body, true);
+
+        $emailToFind = $data['dynamicID']['email'] ?? '';
+
+        $optInEmail = $optInEmailRepository->findById($emailToFind);
+        if (!$optInEmail) {
+            throw new OutOfBoundsException('No Opt-In Email found for ' . $emailToFind);
+        }
 
         $optInEmail->email = $data['email'] ?? '';
         $errors = $validator->validate($optInEmail);
@@ -87,15 +106,23 @@ class OptInEmailController
         return new JsonResponse($params);
     }
 
-    #[Route('/{optInEmail}', methods: ['DELETE'])]
-    #[IsGranted('OPTIN_EMAIL_DELETE', subject: 'optInEmail')]
-    #[ParamConverter('optInEmail', class: OptInEmail::class, converter: 'app.optInEmail')]
+    #[Route('/delete', methods: ['DELETE'])]
+    #[IsGranted('OPTIN_EMAIL_DELETE')]
     public function delete(
-        OptInEmail           $optInEmail,
+        Request           $request,
         OptInEmailRepository $optInEmailRepository
     ): Response
     {
+        $body = $request->getContent();
+        $data = json_decode($body, true);
+
+        $email = $data['email'] ?? '';
+
+        $optInEmail = $optInEmailRepository->findById($email);
+        if (!$optInEmail) {
+            throw new OutOfBoundsException('No Opt-In Email found for ' . $email);
+        }
         $optInEmailRepository->delete($optInEmail);
-        return new JsonResponse('Email deleted successfully!');
+        return new JsonResponse('Record deleted successfully!');
     }
 }
